#853: page deleting updated.

+++ Modules/BetterCms.Module.Pages/Command/Page/DeletePage/DeletePageCommand.cs
@@ -47,8 +47,8 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
         /// <param name="redirectService">The redirect service.</param>
         /// <param name="sitemapService">The sitemap service.</param>
         /// <param name="urlService">The URL service.</param>
-        public DeletePageCommand(IRedirectService redirectService,
-            ISitemapService sitemapService, IUrlService urlService, ICmsConfiguration cmsConfiguration)
+        /// <param name="cmsConfiguration">The CMS configuration.</param>
+        public DeletePageCommand(IRedirectService redirectService, ISitemapService sitemapService, IUrlService urlService, ICmsConfiguration cmsConfiguration)
         {
             this.redirectService = redirectService;
             this.sitemapService = sitemapService;
@@ -61,6 +61,9 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
         /// </summary>
         /// <param name="request">The request.</param>
         /// <returns></returns>
+        /// <exception cref="ConcurrentDataException"></exception>
+        /// <exception cref="System.ComponentModel.DataAnnotations.ValidationException">
+        /// </exception>
         public virtual bool Execute(DeletePageViewModel request)
         {
             var page = Repository.First<PageProperties>(request.PageId);
@@ -78,28 +81,29 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
 
             request.RedirectUrl = urlService.FixUrl(request.RedirectUrl);
 
-            IList<SitemapNode> sitemapNodes = null;
-
             if (request.UpdateSitemap)
             {
                 AccessControlService.DemandAccess(Context.Principal, RootModuleConstants.UserRoles.EditContent);
             }
 
-            sitemapNodes = sitemapService.GetNodesByPage(page);
+            var sitemaps = new Dictionary<Models.Sitemap, bool>();
+            var sitemapNodes = sitemapService.GetNodesByPage(page);
             if (request.UpdateSitemap)
             {
-                if (cmsConfiguration.Security.AccessControlEnabled)
-                {
-                    var sitemaps = sitemapNodes.Select(node => node.Sitemap).Distinct().ToList();
-                    sitemaps.ForEach(sitemap => AccessControlService.DemandAccess(sitemap, Context.Principal, AccessLevel.ReadWrite));
-                    sitemaps.ForEach(sitemap => sitemapService.ArchiveSitemap(sitemap.Id));
-                }
+                sitemapNodes.Select(node => node.Sitemap)
+                            .Distinct()
+                            .ToList()
+                            .ForEach(
+                                sitemap =>
+                                sitemaps.Add(
+                                    sitemap,
+                                    !cmsConfiguration.Security.AccessControlEnabled || AccessControlService.GetAccessLevel(sitemap, Context.Principal) == AccessLevel.ReadWrite));
 
                 foreach (var node in sitemapNodes)
                 {
-                    if (node.ChildNodes.Count > 0)
+                    if (sitemaps[node.Sitemap] && node.ChildNodes.Count > 0)
                     {
-                        var logMessage = string.Format("Sitemap node {0} has {1} child nodes.", node.Id, node.ChildNodes.Count);
+                        var logMessage = string.Format("In {0} sitemap node {1} has {2} child nodes.", node.Sitemap.Id, node.Id, node.ChildNodes.Count);
                         throw new ValidationException(() => PagesGlobalization.DeletePageCommand_SitemapNodeHasChildNodes_Message, logMessage);
                     }
                 }
@@ -112,32 +116,27 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
             if (sitemapNodes != null)
             {
                 // Archive sitemaps before update.
-                sitemapNodes.Select(node => node.Sitemap).Distinct().ToList()
-                    .ForEach(sitemap => sitemapService.ArchiveSitemap(sitemap.Id));
-
-                if (request.UpdateSitemap)
+                sitemaps.Select(pair => pair.Key).ToList().ForEach(sitemap => sitemapService.ArchiveSitemap(sitemap.Id));
+                foreach (var node in sitemapNodes)
                 {
-                    // Delete sitemapNodes.
-                    foreach (var node in sitemapNodes)
+                    if (!node.IsDeleted)
                     {
-                        if (!node.IsDeleted)
+                        if (request.UpdateSitemap && sitemaps[node.Sitemap])
                         {
+                            // Delete sitemap node.
                             sitemapService.DeleteNode(node, ref deletedNodes);
                         }
-                    }
-                }
-                else
-                {
-                    // Unlink sitemap nodes.
-                    foreach (var node in sitemapNodes)
-                    {
-                        if (node.Page != null && node.Page.Id == page.Id)
+                        else
                         {
-                            node.Page = null;
-                            node.Url = page.PageUrl;
-                            node.UrlHash = page.PageUrlHash;
-                            Repository.Save(node);
-                            updatedNodes.Add(node);
+                            // Unlink sitemap node.
+                            if (node.Page != null && node.Page.Id == page.Id)
+                            {
+                                node.Page = null;
+                                node.Url = page.PageUrl;
+                                node.UrlHash = page.PageUrlHash;
+                                Repository.Save(node);
+                                updatedNodes.Add(node);
+                            }
                         }
                     }
                 }
@@ -189,6 +188,7 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
                     Repository.Delete(pageTag);
                 }
             }
+
             if (page.PageContents != null)
             {
                 foreach (var pageContent in page.PageContents)
@@ -196,6 +196,7 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
                     Repository.Delete(pageContent);
                 }
             }
+
             if (page.Options != null)
             {
                 foreach (var option in page.Options)
@@ -203,11 +204,13 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
                     Repository.Delete(option);
                 }
             }
+
             if (page.AccessRules != null)
             {
                 var rules = page.AccessRules.ToList();
                 rules.ForEach(page.RemoveRule);
             }
+
             if (page.MasterPages != null)
             {
                 foreach (var master in page.MasterPages)
@@ -248,6 +251,12 @@ namespace BetterCms.Module.Pages.Command.Page.DeletePage
 
             // Notifying, that page is deleted.
             Events.PageEvents.Instance.OnPageDeleted(page);
+
+            if (sitemaps.Any(tuple => !tuple.Value))
+            {
+                // Some sitemaps where skipped, because user has no permission to edit.
+                Context.Messages.AddSuccess(PagesGlobalization.DeletePage_SitemapSkipped_Message);
+            }
 
             return true;
         }
